update list entity pages to make them consistent
