+++ studio/src/pages/formats/index.js
@@ -1,10 +1,44 @@
 import React from 'react';
 import FormatList from './components/FormatList';
 import { Space, Button, Row } from 'antd';
-import { Link } from 'react-router-dom';
+import { Link, useLocation } from 'react-router-dom';
+import { useDispatch, useSelector } from 'react-redux';
+import { getFormats } from '../../actions/formats';
+import deepEqual from 'deep-equal';
 
 function Formats({ permission }) {
   const { actions } = permission;
+  const dispatch = useDispatch();
+  const query = new URLSearchParams(useLocation().search);
+  const [filters, setFilters] = React.useState({
+    page: 1,
+    limit: 20,
+  });
+  query.set('page', filters.page);
+  window.history.replaceState({}, '', `${window.PUBLIC_URL}${useLocation().pathname}?${query}`);
+  const { formats, total, loading } = useSelector((state) => {
+    const node = state.formats.req.find((item) => {
+      return deepEqual(item.query, filters);
+    });
+
+    if (node)
+      return {
+        formats: node.data.map((element) => state.formats.details[element]),
+        total: node.total,
+        loading: state.formats.loading,
+      };
+    return { formats: [], total: 0, loading: state.formats.loading };
+  });
+
+  React.useEffect(() => {
+    fetchFormats();
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, [filters]);
+
+  const fetchFormats = () => {
+    dispatch(getFormats(filters));
+  };
+
   return (
     <Space direction="vertical">
       <Row justify="end">
@@ -18,7 +52,13 @@ function Formats({ permission }) {
         </Link>
       </Row>
 
-      <FormatList actions={actions} />
+      <FormatList
+        actions={actions}
+        data={{ formats, total, loading }}
+        filters={filters}
+        setFilters={setFilters}
+        fetchFormats={fetchFormats}
+      />
     </Space>
   );
 }
