+++ studio/src/pages/policies/index.js
@@ -1,13 +1,44 @@
 import React from 'react';
 import { Space, Button, Row } from 'antd';
-import { Link } from 'react-router-dom';
+import { Link, useLocation } from 'react-router-dom';
 import PolicyList from './components/PolicyList';
 import getUserPermission from '../../utils/getUserPermission';
-import { useSelector } from 'react-redux';
-
+import { useDispatch, useSelector } from 'react-redux';
+import { getPolicies } from '../../actions/policies';
+import deepEqual from 'deep-equal';
 function Policies() {
   const spaces = useSelector(({ spaces }) => spaces);
   const actions = getUserPermission({ resource: 'policies', action: 'get', spaces });
+  const dispatch = useDispatch();
+  const query = new URLSearchParams(useLocation().search);
+  const [filters, setFilters] = React.useState({
+    page: 1,
+    limit: 20,
+  });
+  query.set('page', filters.page);
+  window.history.replaceState({}, '', `${window.PUBLIC_URL}${useLocation().pathname}?${query}`);
+  const { policies, total, loading } = useSelector((state) => {
+    const node = state.policies.req.find((item) => {
+      return deepEqual(item.query, filters);
+    });
+
+    if (node)
+      return {
+        policies: node.data.map((element) => state.policies.details[element]),
+        total: node.total,
+        loading: state.policies.loading,
+      };
+    return { policies: [], total: 0, loading: state.policies.loading };
+  });
+
+  React.useEffect(() => {
+    fetchPolicies();
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, [filters]);
+
+  const fetchPolicies = () => {
+    dispatch(getPolicies(filters));
+  };
 
   return (
     <Space direction="vertical">
@@ -22,7 +53,13 @@ function Policies() {
         </Link>
       </Row>
 
-      <PolicyList actions={actions} />
+      <PolicyList
+        actions={actions}
+        data={{ policies, total, loading }}
+        filters={filters}
+        setFilters={setFilters}
+        fetchPolicies={fetchPolicies}
+      />
     </Space>
   );
 }
